webmidi/midioutputmap.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="UTF-8">
    <title>Web MIDI API Test: obtain & list MIDI output ports</title>
    <link rel="author" title="Ryoya KAWAI" href="mailto:ryoya.kawai@gmail.com">
    <link rel="help" href="http://webaudio.github.io/web-midi-api/#midioutputmap-interface">
    <link rel="match" href="reference/REFERENCE-FILE-NAME">
    <meta name="assert" content="Return true when test success in outputs().">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <style type="text/css">
    /* ADD STYLE BLOCK HERE (PREFERRABLE TO INLINE STYLES) */
    body { font-family:arial;}
    </style>
  </head>
  <body>
    <p>Success when be obtained OUTPUT MIDI devce and its properties are correct.</p>
    <div id="warning" style="color:red;"></div>

    <div id="log"></div>

    <!-- PAGE CONTENT -->
    <script type="text/javascript">
    var MIDI;
    var timerId={ };
    var getPort={ };

    var test_midiOutputMap = async_test("MIDIOutputMap Interface Test");

    navigator.requestMIDIAccess().then( successCallback, errorCallback );
    function successCallback(access) {
        MIDI = access;
        test_midiOutputMap.step(function() {
            var outputs=[];
            if(typeof MIDI.outputs==="function") {
                document.getElementById("warning").innerHTML= "This browser seems the implementation of the API, Web MIDI API, is followed old spec.";
                outputs=MIDI.outputs();
            } else {
                var outputIterator=MIDI.outputs.values();
                for(var o=outputIterator.next(); !o.done; o=outputIterator.next()) {
                    outputs.push(o.value);
                }
                console.log("[OutputDevices] ", outputs);
                if(outputs.length==0) {
                    document.getElementById("warning").innerHTML="Your Computer looks no <b><u>MIDI OUTPUT</u></b> devices are pluged-in.<br> Please plug-in at least one <b><u>MIDI OUTPUT</u></b> device before run this Test.";
                    assert_false(true);
                    test_midiOutputMap.done();
                } else {
                    test(function() {
                        assert_true(MIDI.outputs.constructor.name=="MIDIOutputMap");
                    }, "MIDIOutputMap interface name test.");
                    test(function() {
                        assert_true(typeof MIDI.outputs.size=="number");
                    }, "MIDIOutputMap interface attribute test.");
                    test(function() {
                        assert_true(typeof MIDI.outputs.keys=="function");
                        assert_true(typeof MIDI.outputs.entries=="function");
                        assert_true(typeof MIDI.outputs.values=="function");
                        assert_true(typeof MIDI.outputs.get=="function");
                        assert_true(typeof MIDI.outputs.has=="function");
                    }, "MIDIOutputMap interface method test.");
                    assert_true(outputs instanceof Object);
                    for( var j=0; j<outputs.length; j++ ) {
                        test(function() {
                            assert_true( typeof outputs[j]["id"]=="string", "[Type of \"id\": String] " );
                            assert_true( typeof outputs[j]["manufactuerer"]=="string" || outputs[j]["manufactuerer"]==null, "[Type of \"manufactuerer\": String or null] " );
                            assert_true( typeof outputs[j]["name"]=="string" || outputs[j]["name"]==null, "[Type of \"name\": String or null] " );
                            assert_in_array( outputs[j]["type"], ["input", "output"], "[Type of \"type\": input or output] " );
                            assert_true( typeof outputs[j]["version"]=="string" || outputs[j]["version"]==null, "[Type of \"version\": String or null] " );
                            assert_true(outputs[j].constructor.name=="MIDIOutput");
                        }, "MIDIPort interface attribute test.");
                        test(function() {
                            assert_true( typeof outputs[j].send=="function");
                        }, "MIDIPort interface MIDIOutput method test.");
                    }
                    test_midiOutputMap.done();
                }
            }
        });
    }
    function errorCallback(message) {
        test_midiOutputMap.step(function() {
            console.log(message);
            assert_true(false);
            test_midiOutputMap.done();
        });
    }


    </script>
  </body>
</html>
